[Chore] adapter package 묶음
