+++ presentation/src/main/java/com/movingmaker/presentation/view/main/gatherdiary/adapter/CommentDiaryCommentYetAdapter.kt
@@ -1,4 +1,4 @@
-package com.movingmaker.presentation.view.main.gatherdiary
+package com.movingmaker.presentation.view.main.gatherdiary.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.movingmaker.domain.model.response.Diary
 import com.movingmaker.presentation.databinding.RvItemCommentDiaryCommentYetBinding
+import com.movingmaker.presentation.view.main.gatherdiary.CommentDiaryDetailViewHolders
 
 class CommentDiaryCommentYetAdapter :
     ListAdapter<Diary, CommentDiaryDetailViewHolders.CommentDiaryCommentYetViewHolder>(
