+++ presentation/src/main/java/com/movingmaker/presentation/view/main/gatherdiary/adapter/CommentDiaryEmptyCommentAdapter.kt
@@ -1,10 +1,11 @@
-package com.movingmaker.presentation.view.main.gatherdiary
+package com.movingmaker.presentation.view.main.gatherdiary.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import com.movingmaker.presentation.databinding.RvItemCommentDiaryEmptyCommentBinding
+import com.movingmaker.presentation.view.main.gatherdiary.CommentDiaryDetailViewHolders
 
 class CommentDiaryEmptyCommentAdapter() :
     ListAdapter<String, CommentDiaryDetailViewHolders.CommentDiaryEmptyCommentViewHolder>(
